% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R, R/mvnpdfoptim.R, R/mvnpdfsmart.R
\name{mvnpdf}
\alias{mvnpdf}
\alias{mvnpdfoptim}
\alias{mvnpdfsmart}
\title{mvnpdf}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)),
  Log = TRUE)

mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)),
  Log = TRUE)
}
\arguments{
\item{x}{your dataframe}

\item{mean}{vector of mean}

\item{varcovM}{variance/covariance matrix}

\item{Log}{if Lof == TRUE, it's return logarythm value}

\item{x}{the vector x}

\item{mean}{mean by colonn}

\item{varcovM}{variance/covariance}

\item{Log}{lograryhtm value}
}
\value{
list of vector, matrix x and the value of density
}
\description{
mvnpdf

mvnpdfoptim

mvnpdfsmart
}
\examples{
x <- 1
}
