% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdfoptim.R
\name{plot.mvnpdf}
\alias{plot.mvnpdf}
\title{x <- 1
plot.mvnpdf}
\usage{
\method{plot}{mvnpdf}(x, ...)
}
\arguments{
\item{x}{vector x}

\item{y}{vector y}
}
\value{
only a plot
}
\description{
x <- 1
plot.mvnpdf
}
\examples{
x <- 1
}
