% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{mcvpdf}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{your dataframe}

\item{mean}{vector of mean}

\item{varcovM}{variance/covariance matrix}

\item{Log}{if Lof == TRUE, it's return logarythm value}
}
\value{
list of vector, matrix x and the value of density
}
\description{
mcvpdf
}
